fixing a few flows - fixing tests
